Add battery voltage sensing.
I have no idea if this works, I developed this commit without access
to hardware. I don't know how long an ADC conversion takes, so it's
set up as an interrupt. There's a state machine that iterates through
all of the analog values to be read, and handles them. There is no
current way for the ISR to tell the main thread that the conversion
has finished, so that needs to be worked out once CAN is working.