Make UART work, we can now communicate with XBEEs
Successfully talked to an XBEE over UART. Modifies the main function
for a very basic uart example. UART module currently only runs over
polling, not interrupt mode. This is solely for ease of development
use, we should at some point in the future break that out into an
interrupt handler.