Create Shark of the Sky Controller file pair
sotscon (shark of the sky controller) takes the CAN messages coming in
from the CAN bus, and updates an internal model based on those messages.
Currently the only thing the model tracks is how long it's been since
we've heard from every board on the bus, and the state of the injector
and vent valve.

This commit also creates a report_error function, since there are
several exception cases in sotscon that need to be reported somehow. The
report_error function doesn't really /do/ anything, but that's fine.
It's there. Ready to do stuff. Just later.

This commit has a major TODO, it does not handle
MSG_GENERAL_BOARD_STATUS CAN messages. It needs to treat those messages
as either a heartbeat (if ERROR_CODE is E_NOMINAL), or an error message
(if ERROR_CODE is something else). So that needs to happen. Otherwise it
should be fine.

The other required thing is adding the sotscon functions in the main
loop somehow. This branch won't be merged into master until a full event
loop has been written. There are still three major components left
before getting that working:

1. Radio IO handler
2. CAN output manager
3. Putting all that together in main()