Setup a full event loop in main
Creates sotscon_sender, a module whose whole goal in life is to send CAN
messages to change the state of the vent and injector valves whenever
they're the wrong state.

This commit updates canlib to point at a version that includes the
receive can buffer ring buffer module. This module is not merged into
master, but a pull request has been submitted. When that pull request is
accepted, I'll update this repo to point at the new version.

This commit also creates a couple of modules that aren't implemented,
but the functions are declared so that I can put it in the main loop,
where it'll eventually be useful. An example of this is radio_handler,
which will someday be implemented to handle incoming radio packets, but
currently just fakes knowing what state the valves should be in.

I also create a micros function, which will someday be used for CAN
message timestamps (sotscon_sender sends CAN messages, so it needs
microsecond timestamps). I haven't implemented it, I don't know how I
will someday do so, but someday it'll be a thing.