Add micros() function, and update to latest canlib
the function in pic18_time.c has a bunch of comment before it. That
should explain everything.

I don't care that the micros funciton is accurate in the short term
(accurate over a few seconds is important, but not less than that). What
is critical is that it is monotonic. It should never be possible (except
in cases where the 32 bit return value overflows) that you call micros()
twice and the second value is lower than the first. I'm pretty sure this
does that.