Add scaling factors for current measurement.
The math is explained in a comment. Long story short, we had to divide a
number by 15. If you wanted to be really cute, you could do some "divide
by sixteen and add it divided by 32 and only get a little bit of error
across this range", but I'm gonna assume that this is good.