Create asynchronous UART rx and tx buffering
Now you can write multiple bytes into the UART module, and the calls to
uart_transmit don't block. This enables the TXIF and RXIF uart
interrupts so that bytes are sent and received as quickly as possible,
not bound by the speed of the application loop. This also moves the uart
initialization into uart.c, out of init.c. Really, we probably shouldn't
have init.c, but that's a problem for another time.

THIS CODE IS NOT TESTED. I know that it compiles, I do not know whether
it works. I don't have the hardware with me here in California to test
whether it works, I promise that that's first on my list to check when I
get back to Waterloo.