Add radio handling code, radio stuff should be done
Added the ability to send error messages over radio, which happens at
most every second. Also added the ability for the operators to request
the state from the rocket (by sending '}'), at which point the rocket
will respond (using data sourced from the sotscon module).

All of this is completely untested, so it needs to be that'd. Also, I
ran all committed files through astyle, so if you're wondering why there
are some pointless whitespace changes, that's why.