Add debug LED blinking to represent stuff
Red LED blinks as long as the board is alive, and it double blinks if
the bus is powered. White LED blinks if the vent valve is supposed to be
open (not if it is, if it's supposed to). Blue LED blinks if injector
valve is supposed to be open.